### 2022-10-09 #4834 @editorialbot
[PRE REVIEW]: PyKronecker: A Python Library for the Efficient Manipulation of Kronecker Products and Related Structures
**Submitting author:** @nickelnine37 (<a href="http://orcid.org/0000-0002-5553-8145">Edward Antonian</a>)
**Repository:** https://github.com/nickelnine37/pykronecker
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.1
