### 2022-10-09 4d8141b
doi bib and small paper updates
