+++ paper.bib
@@ -7,6 +7,7 @@ volume={29},
 number={8},
 pages={3374-3387},
 doi={10.1109/TNNLS.2017.2727545}
+url = {https://doi.org/10.1109/TNNLS.2017.2727545},
 }
 
 @article{loan2000,
@@ -18,8 +19,8 @@ pages = {85-100},
 year = {2000},
 note = {Numerical Analysis 2000. Vol. III: Linear Algebra},
 issn = {0377-0427},
-doi = {https://doi.org/10.1016/S0377-0427(00)00393-9},
-url = {https://www.sciencedirect.com/science/article/pii/S0377042700003939},
+doi = {10.1016/S0377-0427(00)00393-9},
+url = {https://doi.org/10.1016/S0377-0427(00)00393-9},
 author = {Charles F.Van Loan},
 }
 
@@ -68,7 +69,8 @@ volume = {11},
 pages = {100361},
 year = {2020},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2019.100361},
+doi = {10.1016/j.softx.2019.100361},
+url={https://doi.org/10.1016/j.softx.2019.100361},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711019301086},
 author = {Matteo Ravasi and Ivan Vasconcelos},
 keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear operator},
@@ -77,9 +79,10 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
 @inproceedings{Stock2020,
   title={A general-purpose toolbox for efficient Kronecker-based learning},
   author={Michiel Stock and Tapio Pahikkala and Antti Airola and Bernard De Baets},
+  doi={10.21105/jcon.00015},
+  url={https://doi.org/10.21105/jcon.00015},
   year={2020}
 }
-
 
 
 @article{baydin2018,
@@ -89,6 +92,8 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
   volume={18},
   pages={1--43},
   year={2018},
+  doi={10.48550/arXiv.1502.05767},
+  url={https://doi.org/10.48550/arXiv.1502.05767},
   publisher={Microtome Publishing}
 }
 
@@ -109,7 +114,8 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
   pages={65--98},
   year={2017},
   publisher={SIAM},
-  url={https://doi.org/10.1137/141000671}
+  doi={10.1137/141000671},
+  url={https://doi.org/10.1137/141000671},
 }
 
 @inproceedings{shewchuk1994,
@@ -122,6 +128,8 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
    author = {W E Roth},
    journal = {Bulletin of the American Mathematical Society},
    title = {On Direct Product Matrices},
+   doi={10.2307/3609497},
+   url={https://doi.org/10.2307/3609497},
    year = {1934},
 }
 
@@ -131,7 +139,8 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
    issue = {123},
    title = {Efficient Computer Manipulation of Tensor Products with Applications to Multidimensional Approximation},
    volume = {27},
-   url = {https://www.ams.org/journal-terms-of-use},
+   doi={10.1090/s0025-5718-1973-0395196-6},
+   url={https://doi.org/10.1090/s0025-5718-1973-0395196-6},
    year = {1973},
 }
 
@@ -144,12 +153,15 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
    title = {Efficient Computer Manipulation of Tensor Products},
    volume = {5},
    year = {1979},
[…]
