+++ paper.bib
@@ -6,7 +6,7 @@ year={2018},
 volume={29},
 number={8},
 pages={3374-3387},
-doi={10.1109/TNNLS.2017.2727545}
+doi={10.1109/TNNLS.2017.2727545},
 url = {https://doi.org/10.1109/TNNLS.2017.2727545},
 }
 
@@ -27,7 +27,7 @@ author = {Charles F.Van Loan},
 @misc{Schacke2013,
     author = {Kathrin Schäcke},
     title = {On the Kronecker product},
-    year = {2013}
+    year = {2013},
 }
 
 
@@ -81,7 +81,7 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
   author={Michiel Stock and Tapio Pahikkala and Antti Airola and Bernard De Baets},
   doi={10.21105/jcon.00015},
   url={https://doi.org/10.21105/jcon.00015},
-  year={2020}
+  year={2020},
 }
 
 
@@ -94,7 +94,7 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
   year={2018},
   doi={10.48550/arXiv.1502.05767},
   url={https://doi.org/10.48550/arXiv.1502.05767},
-  publisher={Microtome Publishing}
+  publisher={Microtome Publishing},
 }
 
 @inproceedings{okuta2017,
