+++ paper.bib
@@ -121,7 +121,8 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
 @inproceedings{shewchuk1994,
   title={An Introduction to the Conjugate Gradient Method Without the Agonizing Pain},
   author={Jonathan Richard Shewchuk},
-  year={1994}
+  year={1994},
+  url={https://www.cs.cmu.edu/~quake-papers/painless-conjugate-gradient.pdf},
 }
 
 @article{Roth1934,
