### 2022-10-13 6dd6f3d
small updates to docs
