### 2022-10-14 #4834 @Kevin-Mattheus-Moerman
@dfm @nicoguaro @likask @lizarett @JulianKarlBauer would be interested in reviewing this work for JOSS? Also let me know if I am relying on your review efforts too often :) thank you so much for the help some of you have provided to us in the past already!!!!
