### 2022-10-18 #4834 @nickelnine37
Thanks @Kevin-Mattheus-Moerman, and thank you @nicoguaro, @JulianKarlBauer for agreeing to review. @mikechantler, @garethpeters78 and I all look forward to your feedback.

One thought I have is that, in order to reproduce the performance claims in the paper, some config is required, especially to run on a GPU. I don't think Binder supports GPUs, so perhaps the best option would be to use Google colab, or maybe just provide a Docker file to run in GPU-capable cloud environment. Let me know if you have any thoughts on this.
