### 2022-10-19 61bd532
fixed small inacuracy in KroneckerDiag repr
