### 2022-11-11 8cd5ce6
Make name "Kronecker" uppercase in reference titles
