+++ paper.bib
@@ -1,7 +1,7 @@
 @article{Airola2018,
 author={Airola, Antti and Pahikkala, Tapio},
 journal={IEEE Transactions on Neural Networks and Learning Systems},
-title={Fast Kronecker Product Kernel Methods via Generalized Vec Trick},
+title={Fast {K}ronecker Product Kernel Methods via Generalized Vec Trick},
 year={2018},
 volume={29},
 number={8},
@@ -11,7 +11,7 @@ url = {https://doi.org/10.1109/TNNLS.2017.2727545},
 }
 
 @article{loan2000,
-title = {The ubiquitous Kronecker product},
+title = {The ubiquitous {K}ronecker product},
 journal = {Journal of Computational and Applied Mathematics},
 volume = {123},
 number = {1},
@@ -26,7 +26,7 @@ author = {Charles F.Van Loan},
 
 @misc{Schacke2013,
     author = {Kathrin Schäcke},
-    title = {On the Kronecker product},
+    title = {On the {K}ronecker product},
     year = {2013},
 }
 
@@ -77,7 +77,7 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
 }
 
 @inproceedings{Stock2020,
-  title={A general-purpose toolbox for efficient Kronecker-based learning},
+  title={A general-purpose toolbox for efficient {K}ronecker-based learning},
   author={Michiel Stock and Tapio Pahikkala and Antti Airola and Bernard De Baets},
   doi={10.21105/jcon.00015},
   url={https://doi.org/10.21105/jcon.00015},
@@ -184,7 +184,7 @@ keywords = {Python, Linear algebra, Inverse problems, Optimization, Linear opera
 
 @article{Fackler2019,
 author = {Fackler, Paul L.},
-title = {Algorithm 993: Efficient Computation with Kronecker Products},
+title = {Algorithm 993: Efficient Computation with {K}ronecker Products},
 year = {2019},
 issue_date = {June 2019},
 publisher = {Association for Computing Machinery},
