### 2022-11-11 #4900 @JulianKarlBauer
# Review

### General checks
General checks are passed.

### Functionality
Kronecker product of two random matrices of homogeneous dimension have been calculated by `pykronecker.KroneckerProduct([np.random.rand(n, n), np.random.rand(n, n)])` with values of `n` up to 150 in this [script](https://github.com/JulianKarlBauer/pykronecker/blob/review/review/performance_example_01.py).
Equality check with `np.kron` and `pylops.Kronecker` pass.

Performance claims of the paper can not be reproduced due to limited hardware performance.
However, [`perfplot`](https://github.com/nschloe/perfplot) has been used in this [script](https://github.com/JulianKarlBauer/pykronecker/blob/review/review/performance_example_01.py) to compare `pykronecker` with alternative Python libraries mentioned in the paper. `pykronecker` performs well within these checks.

![image](https://user-images.githubusercontent.com/55383008/201378992-02943d9e-ba8e-49e9-bf50-195f9a0fc3f5.png)

### Documentation
The structure and scope of the documentation are good and appropriate for the size of the package. Installation are given.
Examples are given and cover essential use cases. Tests are automated and pass on multiple Python versions. However, a note on how to run the tests, e.g., placed in directory `tests/README.md` could be an improvement. This file might point towards this [Github-actions file](https://github.com/nickelnine37/pykronecker/blob/main/.github/workflows/tests.yml). Currently, [contributing instructions](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) are missing.

### Software paper

The manuscript is well structured. The formulations are precise and catchy. For readers without direct prior knowledge, a basic understanding is created and the subject matter is well motivated. For example, a reader with knowledge on tensor algebra but without direct prior experience with the Kronkecker product or the Kronecker sum of matrices is picked up on the content. PyKronecker is comprehensibly motivated as an efficient implementation of basic operations with a user-friendly interface. Formal and linguistic requirements are met and the list of references is diverse and appears to be complete.

### Improvements / Recommendations
- See merge-requests and issues mentioned above.
- Please extend community guidelines on contributing and bug reporting.
