### 2022-11-11 600d75b
fixed install instructions for zsh compatibility
