### 2022-11-11 #4900 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman during the first steps of the review process, I created a merge-request with some small changes instead of raising an issue describing what to change. As this merge-request now has been merged, I am listed as contributor to pykronecker. Formally, does this induce a COI? This question might be a bit pedantic, as such a situation is probably common during the review process. But I am a bit in doubt and therefore ask, following https://github.com/openjournals/joss/blob/main/COI.md .

@JulianKarlBauer this would not be seen as a COI. If however the contribution is large enough to warrant co-authorship then it would be.
