### 2022-11-11 #4900 @JulianKarlBauer
> >...
>
> @JulianKarlBauer this would not be seen as a COI. If however the contribution is large enough to warrant co-authorship then it would be.

Thank you for your assessment. The contributions are minimal, so no COI.
