### 2022-11-12 9d1dcd6
made README examples self-contained
