+++ README.md
@@ -42,26 +42,32 @@ from pykronecker import KroneckerProduct
 
 A = np.random.normal(size=(5, 5))
 B = np.random.normal(size=(6, 6))
-C = KroneckerProduct([A, B])
+
+KP = KroneckerProduct([A, B])
 ```
 
-This object can operate on both vectors of shape `(5 * 6, )` and tensors of shape `(5, 6)` using the `@` syntax for matrix multiplication. The returned array will be of the same shape.  
+This object can operate on both vectors of shape `(5 * 6, )` and tensors of shape `(5, 6)` using the `@` syntax for matrix multiplication. The returned array will be of the same shape.
 
 ```python
 x = np.random.normal(size=5 * 6)
-X = np.random.normal(size=(5, 6))
-print(C @ x)
-print(C @ X)
+X = x.reshape(5, 6)
+
+assert np.allclose(KP @ x, (KP @ X).ravel())
 ```
 
 ### KroneckerSum
 
 A `KronekerSum` can be created and used in much the same way.
 ```python
+import numpy as np
 from pykronecker import KroneckerSum
 
-D = KroneckerSum([A, B])
-print(D @ x)
+A = np.random.normal(size=(5, 5))
+B = np.random.normal(size=(6, 6))
+x = np.random.normal(size=5 * 6)
+
+KS = KroneckerSum([A, B])
+print(KS @ x)
 ```
 
 ### KroneckerDiag
@@ -69,22 +75,38 @@ print(D @ x)
 `KroneckerDiag` provides support for diagonal matrices, and can be created by passing a tensor of the appropriate size. This creates, in effect, a matrix with the vectorized tensor along the diagonal. 
 
 ```python
+import numpy as np
 from pykronecker import KroneckerDiag
 
-E = KroneckerDiag(np.random.normal(size=(5, 6)))
-print(E @ x)
+D = np.random.normal(size=(5, 6))
+x = np.random.normal(size=5 * 6)
+
+KD = KroneckerDiag(D)
+print(KD @ x)
 ```
 
 ### KroneckerIdentity
 
-Finally, `KroneckerIdentity` creates the identity matrix, which can be instantiated by passing another operator of the same size, or the shape of tensors it should act on. 
+Finally, `KroneckerIdentity` creates the identity matrix, which can be instantiated by passing another operator of the same size, or the shape of tensors the operator is expected to act on. 
 
 ```python
-from pykronecker import KroneckerIdentity
+import numpy as np
+from pykronecker import KroneckerIdentity, KroneckerDiag
 
-I1 = KroneckerIdentity(like=E)
-I2 = KroneckerIdentity(tensor_shape=(5, 6))
-print(I1 @ x, I2 @ x)
+# create another KroneckerDiag operator
+D = np.random.normal(size=(5, 6))
+KD = KroneckerDiag(D)
+
+# create a KroneckerIdentity by passing `like` parameter
+KI1 = KroneckerIdentity(like=KD)
+
+# create KroneckerIdentity by passing `tensor_shape` parameter
+KI2 = KroneckerIdentity(tensor_shape=(5, 6))
+
+x = np.random.normal(size=5 * 6)
+
+assert np.allclose(KI1 @ x, x)
+assert np.allclose(KI2 @ x, x)
 ```
 
 ## Deriving new operators
@@ -92,56 +114,102 @@ print(I1 @ x, I2 @ x)
 All four of these objects can be added or multiplied together arbitrarily to create new composite operators. In this way, they can be treated similarly to literal NumPy arrays. 
 
 ```python
-F = C @ D + C @ E
[…]
