### 2022-11-12 #4900 @nickelnine37
Thanks @JulianKarlBauer for your time reviewing this. I have now added contributing guidelines here https://github.com/nickelnine37/pykronecker/pull/7
