### 2022-11-12 #4900 @nickelnine37
I have also now added testing instructions in a `tests/README` file. https://github.com/nickelnine37/pykronecker/pull/8
