### 2022-11-12 #4900 @JulianKarlBauer
> Thanks @JulianKarlBauer for your time reviewing this. I have now added contributing guidelines here [nickelnine37/pykronecker#7](https://github.com/nickelnine37/pykronecker/pull/7)

@nickelnine37 Thanks again for adding the contributing guidelines.
@Kevin-Mattheus-Moerman All boxes of my review-checklist are ticked.
