### 2022-11-13 dd39f11
fixed small bug in README example
