### 2022-11-15 493aa09
small improvements to README files
