### 2023-01-16 #4900 @nicoguaro
@danielskatz, I have checked almost all the boxes. I need to go to a different computer to check the tests since they do not work on Windows, which I have on my office computer.

I would also like to play a little bit more... I think I can manage it this week.
