### 2023-01-24 #4900 @Kevin-Mattheus-Moerman
@nickelnine37 both reviewers now recommend acceptance in JOSS so we are good to move on to formal acceptance. There are a couple of steps needed before we can proceed:

- [ ] Can you please archive a copy of the software on ZENODO. You can do this manually or using these automated steps: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
- [ ] Once you have created the ZENODO archive, please report back here with the ZENODO DOI

Please check, and if needed manually edit, the ZENODO archive such that:
- [ ] The ZENODO archive license matches the software license
- [ ] The ZENODO archive title matches the JOSS paper title
- [ ] The author set and order is the same as the JOSS paper
- [ ] You may edit the author descriptions on ZENODO to include ORCID information (recommended, but not required)
- [ ] Please let us know if the version tag has moved on, e.g. this review is currently for `0.1.1`, but if this needs to be changed let us know. The version tag should match one on your repository and the one listed on the ZENODO archive.
