+++ setup.cfg
@@ -4,7 +4,7 @@ description = Tools for performing efficient Kronecker product-based matrix oper
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = nickelnine37
-version=0.1.1
+version=0.1.2
 license = MIT
 license_file = LICENSE
 url=https://github.com/nickelnine37/pykronecker
