### 2023-01-24 #4900 @nickelnine37
@Kevin-Mattheus-Moerman

Thanks a lot. The DOI is 10.5281/zenodo.7566803. I have moved the version tag on to 0.1.2. The code has not changed at all from v0.1.1, but I believe I had to create a new release in order for it to be registered by Zenodo. I have also just pushed a change to correct the UK/USA issue.
